implemented get_date and updated enumerateand select
As part of making a very small change to enumerate I have the Escape key return false now, instead of a descriptive string.

Its cleaner this way and lets the user cancel things smoother now with escape.

Formalized a nested if loop format to handle an error in a series of inputs. Cancel anytime now with tailored processing for each fail, or just pass on the false to cascade the failure up the chain,